==> 00-Preface.R <==
#' # Preface {-}
#' 
#' Since you are reading this book, you are likely
#' 
#' The material in this book started as class slid
#' 
#' Another motivation for writing this book is my 
#' 
#' In this book, we will not work on the advanced 
#' 
#' The book includes the following chapters:
#' 
#' - **[Chapter 1 - Introduction](#Introduction)**
#' 
#' - **[Chapter 2 - Basic Operations](#basicoperat
#' 
#' - **[Chapter 3 - Basic Classes](#BasicObjects)*
#' 
#' - **[Chapter 4 - Data Structure Classes](#DataS
#' 
#' - **[Chapter 5 - Financial Data and Common Oper
#' 
#' - **[Chapter 6 - Importing and Exporting Data f
#' 
#' - **[Chapter 7 - Importing Financial Data from 
#' 
#' - **[Chapter 8 - Programming and Data Analysis 
#' 
#' - **[Chapter 9 - Creating and Saving Figures wi
#' 
#' 
#' - **[Chapter 10 - Financial Econometrics with R
#' 
#' - **[Chapter 11 - Writing Research Scripts](#re
#' 
#' I'm a fan of open source code. All the code use
#' 
#' \vspace{0.5cm}
#' 
#' Book site: [https://sites.google.com/view/pafdr
#' 
#' Personal page: [https://sites.google.com/site/m
#' 
#' Blog _R and Finance_: [https://msperlin.github.
#' 
#' \vspace{0.5cm}
#' 
#' A suggestion, before you read the rest of the b
#' 
#' I hope you enjoy this book and find it useful. 
#' 
#' Good reading!
#' 
#' Marcelo S. Perlin
