==> 01-Introduction.R <==
#' # Introduction {#Introduction}
#' 
#' **In the digital era, information is abundant a
#' 
#' In particular, datasets related to Economics an
#' 
#' Technological advancements were accompanied by 
#' 
#' In this setup, the role of R, a programming lan
#' 
#' 
#' ## What is R
#' 
#' R is a programming language specially designed 
#' 
#' Today, R is almost synonymous to data analysis,
#' 
#' 
#' ## Why Choose R
#' 
#' Learning a new programming language requires a 
#' 
#' First, **R is a mature, stable platform, contin
#' 
#' **Learning R is easy**. My experience in teachi
#' 
#' **The engine of R and the interface of RStudio 
#' 
#' **R is compatible with different operating syst
#' 
#' **R is free!** The main software and all its pa
#' 
#' 
#' ## What Can You Do With R and RStudio?
#' 
#' R is a fairly complete programming language and
#' 
#' * Import, export, process, and store financial 
#' 
#' * Substitute and improve data intensive tasks f
#' 
#' * Develop routines for managing and controlling
#' 
#' * Implementation of various possibilities of em
#' 
#' * Create dynamic _websites_ with the `Shiny` pa
#' 
#' * Create an automated process of developing tec
#' 
#' * Write a technical book with `bookdown`; \inde
#' 
#' * Write and publish a blog about finance with `
#' 
#' Besides the previously highlighted uses, public
#' 
#' 
#' ## Installing R and RStudio
#' 
#' Before going further, let's install the require
#' 
#' 
#' The next screen gives you a choice of mirror to
#' 
#' 
#' The next step involves selecting your operating
#' 
#' 
#' After clicking the link _Download R for Windows
#' 
#' 
#' After clicking the link _base_, the next screen
#' 
#' 
#' 
#' The base installation of R includes its own _GU
#' 
#' The installation of RStudio is simpler than tha
#' 
#' I emphasize that using RStudio is not essential
#' 
#' ## Resources in the Web
#' 
#' **The R community is vivid and engaging**. Ther
#' 
#' Learning and using R can be a social experience
#' 
#' 
#' ## Structure and Organization
#' 
#' This book presents a practical approach to the 
#' 
#' Learning to program in a new language is like l
#' 
#' Throughout the book, every demonstration of cod
#' 
## ------------------------------------------------------------------------
# create a list
x <- list('abc', 1:5, 'dec')

# print list
print(x)

#' 
#' For the previous chunk of code, lines `x <- lis
#' 
#' Code can also be spatially organized using new 
#' 
## ------------------------------------------------------------------------
# create a list
x <- list('abc', 
          1:5, 
		  'dec')

# print list
print(x)

#' 
#' The code also follows a well-defined structure.
#' 
#'
